+++ app_i/src/main/AndroidManifest.xml
@@ -183,7 +183,7 @@
         <!--            </intent-filter>-->
         <!--        </service>-->
         <service
-            android:name=".NotificationListenerService"
+            android:name=".service.NotificationListenerService"
             android:label="@string/app_name"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE"
             android:exported="true">
