+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/NewsNotificationRestarter.java
@@ -1,13 +1,13 @@
 //package com.recoveryrecord.surveyandroid.example;
 //
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_COLLECTION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_CYCLE_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_CYCLE_VALUE_FAILED_RESTART;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_DEVICE_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_STATUS_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_STATUS_VALUE_RESTART;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_TIME;
-//import static com.recoveryrecord.surveyandroid.example.Constants.USER_COLLECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_COLLECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_CYCLE_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_CYCLE_VALUE_FAILED_RESTART;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_DEVICE_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_STATUS_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_STATUS_VALUE_RESTART;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_TIME;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.USER_COLLECTION;
 //
 //import android.annotation.SuppressLint;
 //import android.content.BroadcastReceiver;
