+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/NewsNotificationService.java
@@ -1,51 +1,51 @@
 //package com.recoveryrecord.surveyandroid.example;
 //
-//import static com.recoveryrecord.surveyandroid.example.Constants.COMPARE_RESULT_COLLECTION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.COMPARE_RESULT_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.COMPARE_RESULT_MEDIA;
-//import static com.recoveryrecord.surveyandroid.example.Constants.COMPARE_RESULT_NEW_TITLE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.COMPARE_RESULT_PUBDATE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.DEFAULT_NEWS_CHANNEL_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.DEFAULT_NEWS_NOTIFICATION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.DEFAULT_NEWS_NOTIFICATION_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.DOC_ID_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.GROUP_NEWS_SERVICE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_CHANNEL_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_ID_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_MEDIA_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_COLLECTION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_CYCLE_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_CYCLE_VALUE_SERVICE_PAGE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_DEVICE_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_STATUS_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_STATUS_VALUE_INITIAL;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_SERVICE_TIME;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NOTIFICATION_TYPE_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NOTIFICATION_TYPE_VALUE_NEWS;
-//import static com.recoveryrecord.surveyandroid.example.Constants.NOTIFICATION_TYPE_VALUE_SERVICE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_CLICK;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_COLLECTION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_DEVICE_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_DOC_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_MEDIA;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_NOTI_TIME;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_OPEN_TIME;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_PUBDATE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_RECEIEVE_TIME;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_REMOVE_TIME;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_REMOVE_TYPE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_TITLE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_TYPE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_USER_ID;
-//import static com.recoveryrecord.surveyandroid.example.Constants.SHARE_NOTIFICATION_FIRST_CREATE;
-//import static com.recoveryrecord.surveyandroid.example.Constants.SHARE_PREFERENCE_PUSH_NEWS_MEDIA_LIST_SELECTION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.SHARE_PREFERENCE_USER_NAME;
-//import static com.recoveryrecord.surveyandroid.example.Constants.TEST_USER_COLLECTION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.TRIGGER_BY_KEY;
-//import static com.recoveryrecord.surveyandroid.example.Constants.TRIGGER_BY_VALUE_NOTIFICATION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.USER_COLLECTION;
-//import static com.recoveryrecord.surveyandroid.example.Constants.VIBRATE_EFFECT;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.COMPARE_RESULT_COLLECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.COMPARE_RESULT_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.COMPARE_RESULT_MEDIA;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.COMPARE_RESULT_NEW_TITLE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.COMPARE_RESULT_PUBDATE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.DEFAULT_NEWS_CHANNEL_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.DEFAULT_NEWS_NOTIFICATION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.DEFAULT_NEWS_NOTIFICATION_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.DOC_ID_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.GROUP_NEWS_SERVICE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_CHANNEL_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_ID_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_MEDIA_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_COLLECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_CYCLE_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_CYCLE_VALUE_SERVICE_PAGE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_DEVICE_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_STATUS_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_STATUS_VALUE_INITIAL;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_SERVICE_TIME;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NOTIFICATION_TYPE_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NOTIFICATION_TYPE_VALUE_NEWS;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.NOTIFICATION_TYPE_VALUE_SERVICE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_CLICK;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_COLLECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_DEVICE_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_DOC_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_MEDIA;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_NOTI_TIME;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_OPEN_TIME;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_PUBDATE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_RECEIEVE_TIME;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_REMOVE_TIME;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_REMOVE_TYPE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_TITLE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_TYPE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_USER_ID;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_NOTIFICATION_FIRST_CREATE;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_PREFERENCE_PUSH_NEWS_MEDIA_LIST_SELECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_PREFERENCE_USER_NAME;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.TEST_USER_COLLECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.TRIGGER_BY_KEY;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.TRIGGER_BY_VALUE_NOTIFICATION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.USER_COLLECTION;
+//import static com.recoveryrecord.surveyandroid.example.config.Constants.VIBRATE_EFFECT;
 //
 //import android.annotation.SuppressLint;
 //import android.app.AlarmManager;
@@ -78,7 +78,7 @@
 //import com.recoveryrecord.surveyandroid.example.DbHelper.PushNewsDbHelper;
 //import com.recoveryrecord.surveyandroid.example.activity.NewsModuleActivity;
 //import com.recoveryrecord.surveyandroid.example.sqlite.PushNews;
-//import com.recoveryrecord.surveyandroid.example.ui.MediaType;
+//import com.recoveryrecord.surveyandroid.example.model.MediaType;
 //
 //import java.text.SimpleDateFormat;
 //import java.util.Arrays;
