+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/activity/LoginActivity.kt
@@ -17,9 +17,9 @@ import com.google.firebase.auth.AuthResult
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.auth.GoogleAuthProvider
-import com.recoveryrecord.surveyandroid.example.Constants.SHARE_PREFERENCE_USER_NAME
 import com.recoveryrecord.surveyandroid.example.NewsHybridActivity
 import com.recoveryrecord.surveyandroid.example.R
+import com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_PREFERENCE_USER_ID
 import com.recoveryrecord.surveyandroid.util.showToast
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -67,7 +67,7 @@ class LoginActivity : AppCompatActivity() {
         user?.apply {
             val sharedPrefs = PreferenceManager.getDefaultSharedPreferences(applicationContext)
             val editor = sharedPrefs.edit()
-            editor.putString(SHARE_PREFERENCE_USER_NAME, email?.split("@")?.get(0))
+            editor.putString(SHARE_PREFERENCE_USER_ID, email?.split("@")?.get(0))
             editor.apply()
             startActivity(Intent(this@LoginActivity, NewsHybridActivity::class.java))
             finish()
