+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/activity/NewsModuleActivity.java
@@ -1,51 +1,51 @@
 package com.recoveryrecord.surveyandroid.example.activity;
 
-import static com.recoveryrecord.surveyandroid.example.Constants.MEDIA_COLLECTION;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_COLLECTION;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_CONTENT;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_ID;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_ID_KEY;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_IMAGE;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_MEDIA_KEY;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_PUBDATE;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_TITLE;
-import static com.recoveryrecord.surveyandroid.example.Constants.NEWS_URL;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_BYTE_PER_LINE;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_COLLECTION;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_CONTENT_LENGTH;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_DEVICE_ID;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_DISPLAY_HEIGHT;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_DISPLAY_WIDTH;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_DOC_ID;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_DRAG_NUM;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_DRAG_RECORD;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_FLING_NUM;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_FLING_RECORD;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_FONT_SIZE;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_HAS_IMAGE;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_IMAGE_URL;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_IN_TIME;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_IN_TIME_LONG;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_MEDIA;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_NEWS_ID;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_OUT_TIME;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_PAUSE_COUNT;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_PUBDATE;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_ROW_SPACING;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_SAMPLE_CHECK_ID;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_SHARE;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_TIME_ON_PAGE;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_TIME_SERIES;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_TITLE;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_TRIGGER_BY;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_TRIGGER_BY_NOTIFICATION;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_USER_ID;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_VIEWPORT_NUM;
-import static com.recoveryrecord.surveyandroid.example.Constants.READING_BEHAVIOR_VIEWPORT_RECORD;
-import static com.recoveryrecord.surveyandroid.example.Constants.READ_TOTAL;
-import static com.recoveryrecord.surveyandroid.example.Constants.SHARE_PREFERENCE_TEST_SIZE;
-import static com.recoveryrecord.surveyandroid.example.Constants.SHARE_PREFERENCE_USER_NAME;
-import static com.recoveryrecord.surveyandroid.example.Constants.TRIGGER_BY_KEY;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.MEDIA_COLLECTION;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_COLLECTION;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_CONTENT;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_ID;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_ID_KEY;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_IMAGE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_MEDIA_KEY;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_PUBDATE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_TITLE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_URL;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_BYTE_PER_LINE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_COLLECTION;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_CONTENT_LENGTH;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_DEVICE_ID;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_DISPLAY_HEIGHT;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_DISPLAY_WIDTH;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_DOC_ID;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_DRAG_NUM;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_DRAG_RECORD;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_FLING_NUM;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_FLING_RECORD;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_FONT_SIZE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_HAS_IMAGE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_IMAGE_URL;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_IN_TIME;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_IN_TIME_LONG;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_MEDIA;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_NEWS_ID;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_OUT_TIME;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_PAUSE_COUNT;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_PUBDATE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_ROW_SPACING;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_SAMPLE_CHECK_ID;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_SHARE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_TIME_ON_PAGE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_TIME_SERIES;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_TITLE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_TRIGGER_BY;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_TRIGGER_BY_NOTIFICATION;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_USER_ID;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_VIEWPORT_NUM;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READING_BEHAVIOR_VIEWPORT_RECORD;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.READ_TOTAL;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_PREFERENCE_TEST_SIZE;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_PREFERENCE_USER_ID;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.TRIGGER_BY_KEY;
 
 import android.annotation.SuppressLint;
 import android.app.PendingIntent;
@@ -89,14 +89,14 @@ import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.recoveryrecord.surveyandroid.SimpleGestureListener;
-import com.recoveryrecord.surveyandroid.example.GestureListener;
 import com.recoveryrecord.surveyandroid.example.NewsHybridActivity;
 import com.recoveryrecord.surveyandroid.example.R;
+import com.recoveryrecord.surveyandroid.example.model.MediaType;
 import com.recoveryrecord.surveyandroid.example.model.ReadingBehavior;
 import com.recoveryrecord.surveyandroid.example.receiever.ApplicationSelectorReceiver;
 import com.recoveryrecord.surveyandroid.example.sqlite.DragObj;
 import com.recoveryrecord.surveyandroid.example.sqlite.FlingObj;
-import com.recoveryrecord.surveyandroid.example.ui.MediaType;
+import com.recoveryrecord.surveyandroid.example.ui.GestureListener;
 
 import java.io.InputStream;
 import java.text.ParseException;
@@ -1425,7 +1425,7 @@ public class NewsModuleActivity extends AppCompatActivity implements SimpleGestu
 
         readingBehavior.put(READING_BEHAVIOR_DOC_ID, device_id + " " + myReadingBehavior.getInTimestamp());
         readingBehavior.put(READING_BEHAVIOR_DEVICE_ID, device_id);
-        readingBehavior.put(READING_BEHAVIOR_USER_ID, sharedPrefs.getString(SHARE_PREFERENCE_USER_NAME, "尚未設定實驗編號"));
+        readingBehavior.put(READING_BEHAVIOR_USER_ID, sharedPrefs.getString(SHARE_PREFERENCE_USER_ID, "尚未設定實驗編號"));
         readingBehavior.put(READING_BEHAVIOR_SAMPLE_CHECK_ID, "NA");
         //select esm id
         readingBehavior.put(READING_BEHAVIOR_TRIGGER_BY, myReadingBehavior.getTriggerBy());
@@ -1462,7 +1462,7 @@ public class NewsModuleActivity extends AppCompatActivity implements SimpleGestu
 
         myReadingBehavior.setDocId(device_id + " " + myReadingBehavior.getInTimestamp());
         myReadingBehavior.setDeviceId(device_id);
-        myReadingBehavior.setUserId(sharedPrefs.getString(SHARE_PREFERENCE_USER_NAME, "尚未設定實驗編號"));
+        myReadingBehavior.setUserId(sharedPrefs.getString(SHARE_PREFERENCE_USER_ID, "尚未設定實驗編號"));
         myReadingBehavior.setSelectEsmId("NA");
         myReadingBehavior.setHasImg(has_img);
         myReadingBehavior.setPubdate(0);
