+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/receiever/AppUsageReceiver.java
@@ -1,9 +1,9 @@
 package com.recoveryrecord.surveyandroid.example.receiever;
 
-import static com.recoveryrecord.surveyandroid.example.Constants.SHARE_PREFERENCE_USER_NAME;
-import static com.recoveryrecord.surveyandroid.example.config.Constants.DetectTime;
-import static com.recoveryrecord.surveyandroid.example.config.Constants.SessionID;
-import static com.recoveryrecord.surveyandroid.example.config.Constants.UsingApp;
+import static com.recoveryrecord.surveyandroid.example.config.Config.DetectTime;
+import static com.recoveryrecord.surveyandroid.example.config.Config.SessionID;
+import static com.recoveryrecord.surveyandroid.example.config.Config.UsingApp;
+import static com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_PREFERENCE_USER_ID;
 
 import android.annotation.SuppressLint;
 import android.app.Service;
@@ -81,7 +81,7 @@ AppUsageReceiver extends Service {
             myappusage.setKEY_TIMESTAMP(Timestamp.now().getSeconds());
             myappusage.setKEY_DOC_ID(device_id + " " + time_now);
             myappusage.setKEY_DEVICE_ID(device_id);
-            myappusage.setKEY_USER_ID(sharedPrefs.getString(SHARE_PREFERENCE_USER_NAME, "尚未設定實驗編號"));
+            myappusage.setKEY_USER_ID(sharedPrefs.getString(SHARE_PREFERENCE_USER_ID, "尚未設定實驗編號"));
             myappusage.setKEY_SESSION(SessionID);
             myappusage.setKEY_USING_APP(UsingApp);
             myappusage.setKEY_APPUSAGE(foregroundActivityName);
@@ -126,7 +126,7 @@ AppUsageReceiver extends Service {
         myappusage.setKEY_TIMESTAMP(Timestamp.now().getSeconds());
         myappusage.setKEY_DOC_ID(device_id + " " + time_now);
         myappusage.setKEY_DEVICE_ID(device_id);
-        myappusage.setKEY_USER_ID(sharedPrefs.getString(SHARE_PREFERENCE_USER_NAME, "尚未設定實驗編號"));
+        myappusage.setKEY_USER_ID(sharedPrefs.getString(SHARE_PREFERENCE_USER_ID, "尚未設定實驗編號"));
         myappusage.setKEY_SESSION(SessionID);
         myappusage.setKEY_USING_APP(UsingApp);
         myappusage.setKEY_APPUSAGE(foregroundActivityName);
