+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/service/FirebaseService.kt
@@ -17,37 +17,37 @@ import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
-import com.recoveryrecord.surveyandroid.example.Constants.DEFAULT_NEWS_CHANNEL_ID
-import com.recoveryrecord.surveyandroid.example.Constants.GROUP_NEWS
-import com.recoveryrecord.surveyandroid.example.Constants.NEWS_CHANNEL_ID
-import com.recoveryrecord.surveyandroid.example.Constants.NEWS_ID_KEY
-import com.recoveryrecord.surveyandroid.example.Constants.NEWS_MEDIA_KEY
-import com.recoveryrecord.surveyandroid.example.Constants.NEWS_TITLE_KEY
-import com.recoveryrecord.surveyandroid.example.Constants.NO_VALUE
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_MEDIA_SELECTION
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_COLLECTION
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_DEVICE_ID
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_DOC_ID
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_ID
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_MEDIA
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_NOTI_TIME
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_OPEN_TIME
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_PUBDATE
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_RECEIEVE_TIME
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_REMOVE_TIME
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_REMOVE_TYPE
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_TITLE
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_TYPE
-import com.recoveryrecord.surveyandroid.example.Constants.PUSH_NEWS_USER_ID
-import com.recoveryrecord.surveyandroid.example.Constants.SHARE_PREFERENCE_PUSH_NEWS_MEDIA_LIST_SELECTION
-import com.recoveryrecord.surveyandroid.example.Constants.TRIGGER_BY_KEY
-import com.recoveryrecord.surveyandroid.example.Constants.TRIGGER_BY_VALUE_NOTIFICATION
-import com.recoveryrecord.surveyandroid.example.Constants.UPDATE_TIME
-import com.recoveryrecord.surveyandroid.example.Constants.VIBRATE_EFFECT
 import com.recoveryrecord.surveyandroid.example.R
 import com.recoveryrecord.surveyandroid.example.activity.NewsModuleActivity
+import com.recoveryrecord.surveyandroid.example.config.Constants.DEFAULT_NEWS_CHANNEL_ID
+import com.recoveryrecord.surveyandroid.example.config.Constants.GROUP_NEWS
+import com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_CHANNEL_ID
+import com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_ID_KEY
+import com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_MEDIA_KEY
+import com.recoveryrecord.surveyandroid.example.config.Constants.NEWS_TITLE_KEY
+import com.recoveryrecord.surveyandroid.example.config.Constants.NO_VALUE
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_MEDIA_SELECTION
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_COLLECTION
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_DEVICE_ID
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_DOC_ID
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_ID
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_MEDIA
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_NOTI_TIME
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_OPEN_TIME
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_PUBDATE
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_RECEIEVE_TIME
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_REMOVE_TIME
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_REMOVE_TYPE
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_TITLE
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_TYPE
+import com.recoveryrecord.surveyandroid.example.config.Constants.PUSH_NEWS_USER_ID
+import com.recoveryrecord.surveyandroid.example.config.Constants.SHARE_PREFERENCE_PUSH_NEWS_MEDIA_LIST_SELECTION
+import com.recoveryrecord.surveyandroid.example.config.Constants.TRIGGER_BY_KEY
+import com.recoveryrecord.surveyandroid.example.config.Constants.TRIGGER_BY_VALUE_NOTIFICATION
+import com.recoveryrecord.surveyandroid.example.config.Constants.UPDATE_TIME
+import com.recoveryrecord.surveyandroid.example.config.Constants.VIBRATE_EFFECT
+import com.recoveryrecord.surveyandroid.example.model.MediaType
 import com.recoveryrecord.surveyandroid.example.model.NotificationData
-import com.recoveryrecord.surveyandroid.example.ui.MediaType
 import com.recoveryrecord.surveyandroid.util.insertRemote
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
