+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/ui/GestureListener.java
@@ -1,4 +1,4 @@
-package com.recoveryrecord.surveyandroid.example;
+package com.recoveryrecord.surveyandroid.example.ui;
 
 import android.app.Activity;
 import android.util.Log;
