+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/ui/ItemTouchHelperAdapter.java
@@ -1,4 +1,4 @@
-package com.recoveryrecord.surveyandroid.example;
+package com.recoveryrecord.surveyandroid.example.ui;
 
 public interface ItemTouchHelperAdapter {
     void onItemMove(int fromPosition, int toPosition);
