+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/ui/NewsSubFragment.kt
@@ -14,9 +14,9 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.Query
-import com.recoveryrecord.surveyandroid.example.Constants
 import com.recoveryrecord.surveyandroid.example.R
 import com.recoveryrecord.surveyandroid.example.adapter.NewsRecycleViewAdapter
+import com.recoveryrecord.surveyandroid.example.config.Constants
 import com.recoveryrecord.surveyandroid.example.model.News
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
