+++ app_i/src/main/java/com/recoveryrecord/surveyandroid/example/ui/SimpleItemTouchHelperCallback.java
@@ -1,6 +1,5 @@
-package com.recoveryrecord.surveyandroid.example;
+package com.recoveryrecord.surveyandroid.example.ui;
 
-import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
